Fix website ci bug
