+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./Agentica";
-// CONTEXT
+
 export * from "./context/AgenticaContext";
 export * from "./context/AgenticaOperation";
 export * from "./context/AgenticaOperationCollection";
@@ -9,17 +9,15 @@ export * from "./context/AgenticaTokenUsage";
 export * from "./events/AgenticaCallEvent";
 export * from "./events/AgenticaCancelEvent";
 export * from "./events/AgenticaDescribeEvent";
-// EVENTS
 export * from "./events/AgenticaEvent";
 export * from "./events/AgenticaEventSource";
-
 export * from "./events/AgenticaExecuteEvent";
 export * from "./events/AgenticaRequestEvent";
 export * from "./events/AgenticaResponseEvent";
 export * from "./events/AgenticaSelectEvent";
 export * from "./events/AgenticaTextEvent";
+export * as factory from "./factory";
 
-// FACADE CLASS
 export * from "./functional/assertHttpLlmApplication";
 export * from "./functional/validateHttpLlmApplication";
 export * from "./json/IAgenticaEventJson";
@@ -28,17 +26,18 @@ export * from "./json/IAgenticaOperationSelectionJson";
 export * from "./json/IAgenticaPromptJson";
 
 export * from "./json/IAgenticaTokenUsageJson";
+export * as orchestrate from "./orchestrate";
 export * from "./prompts/AgenticaCancelPrompt";
 export * from "./prompts/AgenticaDescribePrompt";
 export * from "./prompts/AgenticaExecutePrompt";
 export * from "./prompts/AgenticaPrompt";
 export * from "./prompts/AgenticaSelectPrompt";
+
 export * from "./prompts/AgenticaTextPrompt";
-// STRUCTURES
 export * from "./structures/IAgenticaConfig";
 export * from "./structures/IAgenticaController";
 export * from "./structures/IAgenticaExecutor";
-
 export * from "./structures/IAgenticaProps";
 export * from "./structures/IAgenticaSystemPrompt";
+
 export * from "./structures/IAgenticaVendor";
