+++ packages/core/src/orchestrate/ChatGptCompletionMessageUtil.ts
